Add condition live testing to action conditions too
